2021.6.28 pass four instructions about invoke test

+++ ch07/instructions/references/invokeinterface.go
@@ -27,7 +27,7 @@ func (self *INVOKE_INTERFACE) Execute(frame *rtda.Frame) {
 		panic("java.lang.IncompatibleClassChangeError")
 	}
 
-	ref := frame.OperandStack().GetRefFromTop(method.ArgSlotCount() - 1)
+	ref := frame.OperandStack().GetRefFromTop(uint(method.ArgSlotCount() - 1))
 
 	if ref == nil {
 		panic("java.lang.NullPointException")
